examples/google_adk_agent/requirements.txt
# Google ADK Agent Requirements
google-adk
google-generativeai
python-dotenv

# MCP and FastAPI dependencies (should already be installed in main project)
mcp
fastapi
pydantic